Bootstrap and AngularJS Incorporation
Bootstrap was decided to be the front end. AngularJS was decided to be
the MVC.

Bootstrap was decided because I'm familiar with it. It's responsive and
can't go wrong with default blue. AngularJS was decided because of its
Google origins. Obviously, NodeJS could have been used since that is the
webserver; however, I felt Angular is also more dev-friendly.